+++ src/main/java/org/fugerit/java/core/web/servlet/helper/ServletUtils.java
@@ -1,5 +1,7 @@
 package org.fugerit.java.core.web.servlet.helper;
 
+import javax.servlet.http.HttpServletResponse;
+
 import org.fugerit.java.core.function.UnsafeVoid;
 
 import lombok.extern.slf4j.Slf4j;
@@ -9,6 +11,10 @@ public class ServletUtils {
 	
 	private ServletUtils() {}
 
+	public static boolean handleSendSafe( HttpServletResponse resp, int httpCode ) {
+		return handleServletOperationSafe( () -> resp.sendError( httpCode ) );
+	}
+	
 	public static boolean handleServletOperationSafe( UnsafeVoid<Exception> fun ) {
 		boolean ok = false;
 		try {